validateToken 구현

+++ src/main/java/juhongBlog/juhongBlog/service/CategoryService.java
@@ -2,6 +2,7 @@ package juhongBlog.juhongBlog.service;
 
 
 import juhongBlog.juhongBlog.domain.Category;
+import juhongBlog.juhongBlog.jwt.JwtTokenProvider;
 import juhongBlog.juhongBlog.repository.CategoryRepository;
 import juhongBlog.juhongBlog.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,9 +16,13 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final PostRepository postRepository;
+    private final JwtTokenProvider jwtTokenProvider;
 
     // 카테고리 추가
-    public String createCategory(Category category) {
+    public String createCategory(Category category, String token) {
+        if(!jwtTokenProvider.validateToken(token)) {
+            return "토큰 유효하지 않음";
+        }
 
         categoryRepository.save(Category.builder()
                 .name(category.getName())
@@ -32,7 +37,10 @@ public class CategoryService {
     }
 
     // 카테고리 삭제
-    public String deleteCategory(Long categoryid) {
+    public String deleteCategory(Long categoryid, String token) {
+        if(!jwtTokenProvider.validateToken(token)) {
+            return "토큰 유효하지 않음";
+        }
         categoryRepository.deleteById(categoryid);
         return "카테고리가 삭제되었습니다.";
     }
